feat: useRenderTick
