+++ src/ui/views/MainPage.tsx
@@ -1,8 +1,13 @@
 import { Box, Text, useInput, useStdout } from "ink";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 
 import { usePage, ViewPage } from "../../hooks/usePage";
-import { useProcessManager, type Process, type ProcessType } from "../../hooks/useProcessManager";
+import {
+	useProcessManager,
+	type Process,
+	type ProcessType,
+} from "../../hooks/useProcessManager";
+import { useRenderTick } from "../../hooks/useRenderTick";
 import { Colors } from "../../lib/Colors";
 import {
 	ShortcutFooter,
@@ -44,7 +49,6 @@ function getReadinessDisplay(process: Process): {
 function ProcessTable() {
 	const { processesRef, selectedProcessIdx } = useProcessManager();
 	const processes = processesRef.current;
-	const [, forceUpdate] = useState(0);
 	const { stdout } = useStdout();
 
 	const terminalWidth = stdout?.columns ?? 80;
@@ -54,15 +58,6 @@ function ProcessTable() {
 		20,
 		terminalWidth - fixedColumnsWidth - borderAndPadding,
 	);
-
-	// Force re-render every second to update age display
-	useEffect(() => {
-		const interval = setInterval(() => {
-			forceUpdate((prev) => prev + 1);
-		}, 1000);
-
-		return () => clearInterval(interval);
-	}, []);
 
 	return (
 		<Box
@@ -215,6 +210,9 @@ export function MainPage() {
 	const { setPage } = usePage();
 	const [showShortcuts, setShowShortcuts] = useState(false);
 	const { stdout } = useStdout();
+
+	// Force re-render every second to update age display and logs
+	useRenderTick();
 
 	const shortcuts = [
 		"↑/↓ or j/k to navigate",
